requirements.txt
Django==3.2.7
djangorestframework==3.12.4
black==21.8b0
Django==3.2.7
flake8==3.9.2
isort==5.9.3
pylint==2.10.2
Pillow==8.3.2
